app/requirements.txt
click
Flask
psycopg2_binary
python-dotenv
pika